+++ src/main/java/com/vega/protocol/model/LiquidityCommitmentOffset.java
@@ -4,13 +4,10 @@ import com.vega.protocol.constant.PeggedReference;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
-import java.math.BigDecimal;
-import java.math.BigInteger;
-
 @Data
 @Accessors(chain = true)
 public class LiquidityCommitmentOffset {
-    private BigInteger proportion;
-    private BigDecimal offset;
+    private long proportion;
+    private double offset;
     private PeggedReference reference;
 }
